start new project

+++ packages/generator/src/Generators.ts
@@ -1,46 +1,34 @@
-import { Mutation, Options, Query } from '@paljs/types';
+import { Mutation, GeneratorOptions, Query, DMMF } from '@paljs/types';
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
 import { format, Options as PrettierOptions } from 'prettier';
 import pkgDir from 'pkg-dir';
 import { join } from 'path';
-import { DMMF } from '@prisma/client/runtime';
-import { getDMMF, getConfig, getEnvPaths, tryLoadEnvs } from '@prisma/sdk';
+import { getInputType, getEnvPaths, tryLoadEnvs, getDMMF } from '@paljs/utils';
 const projectRoot = pkgDir.sync() || process.cwd();
 
 export class Generators {
-  options: Options = {
+  options: GeneratorOptions = {
     prismaName: 'prisma',
     output: join(projectRoot, 'src/graphql'),
     excludeFields: [],
     excludeModels: [],
+    excludeInputFields: [],
     excludeFieldsByModel: {},
     excludeQueriesAndMutations: [],
     excludeQueriesAndMutationsByModel: {},
   };
 
   isJS?: boolean = false;
+  inputName = 'InputTypes';
 
-  queries: Query[] = [
-    'findUnique',
-    'findFirst',
-    'findMany',
-    'findCount',
-    'aggregate',
-  ];
-  mutations: Mutation[] = [
-    'createOne',
-    'updateOne',
-    'upsertOne',
-    'deleteOne',
-    'updateMany',
-    'deleteMany',
-  ];
+  queries: Query[] = ['findUnique', 'findFirst', 'findMany', 'findCount', 'aggregate'];
+  mutations: Mutation[] = ['createOne', 'updateOne', 'upsertOne', 'deleteOne', 'updateMany', 'deleteMany'];
 
   schemaString: string;
 
   readyDmmf?: DMMF.Document;
 
-  constructor(private schemaPath: string, customOptions?: Partial<Options>) {
+  constructor(private schemaPath: string, customOptions?: Partial<GeneratorOptions>) {
     this.options = { ...this.options, ...customOptions };
     this.isJS = this.options.javaScript;
     this.schemaString = readFileSync(this.schemaPath, 'utf-8');
@@ -54,10 +42,6 @@ export class Generators {
     } else {
       return this.readyDmmf;
     }
-  }
-
-  protected async schemaConfig() {
-    return await getConfig({ datamodel: this.schemaString });
   }
 
   protected async datamodel() {
@@ -76,8 +60,7 @@ export class Generators {
   protected async models() {
     const { schema }: { schema: DMMF.Schema } = await this.dmmf();
     return schema.outputObjectTypes.model.filter(
-      (model) =>
-        !this.options.models || this.options.models.includes(model.name),
+      (model) => !this.options.models || this.options.models.includes(model.name),
     );
   }
 
@@ -137,26 +120,19 @@ export class Generators {
   }
 
   protected excludeFields(model: string) {
-    return this.options.excludeFields.concat(
-      this.options.excludeFieldsByModel[model],
-    );
+    return this.options.excludeFields.concat(this.options.excludeFieldsByModel[model]);
   }
 
   protected disableQueries(model: string) {
     return (
-      this.options.disableQueries ||
-      !!this.options.excludeModels.find(
-        (item) => item.name === model && item.queries,
-      )
+      this.options.disableQueries || !!this.options.excludeModels.find((item) => item.name === model && item.queries)
     );
   }
 
   protected disableMutations(model: string) {
     return (
       this.options.disableMutations ||
-      !!this.options.excludeModels.find(
-        (item) => item.name === model && item.mutations,
-      )
+      !!this.options.excludeModels.find((item) => item.name === model && item.mutations)
     );
   }
 
@@ -169,9 +145,7 @@ export class Generators {
   }
 
   protected excludedOperations(model: string) {
-    return this.options.excludeQueriesAndMutations.concat(
-      this.options.excludeQueriesAndMutationsByModel[model] ?? [],
-    );
+    return this.options.excludeQueriesAndMutations.concat(this.options.excludeQueriesAndMutationsByModel[model] ?? []);
   }
 
   protected mkdir(path: string) {
@@ -202,20 +176,71 @@ export class Generators {
     }
   }
 
+  async generateSDLInputsString() {
+    const { schema }: { schema: DMMF.Schema } = await this.dmmf();
+    const fileContent: string[] = ['scalar DateTime', 'type BatchPayload {', 'count: Int!', '}', ''];
+    if (schema) {
+      const enums = [...schema.enumTypes.prisma];
+      if (schema.enumTypes.model) enums.push(...schema.enumTypes.model);
+      enums.forEach((item) => {
+        fileContent.push(`enum ${item.name} {`);
+        item.values.forEach((item2) => {
+          fileContent.push(item2);
+        });
+        fileContent.push('}', '');
+      });
+      const inputObjectTypes = [...schema.inputObjectTypes.prisma];
+      if (schema.inputObjectTypes.model) inputObjectTypes.push(...schema.inputObjectTypes.model);
+
+      inputObjectTypes.forEach((input) => {
+        const inputFields =
+          typeof this.options?.filterInputs === 'function' ? this.options.filterInputs(input) : input.fields;
+        if (inputFields.length > 0) {
+          fileContent.push(`input ${input.name} {`, '');
+          inputFields
+            .filter((field) => !this.options?.excludeInputFields?.includes(field.name))
+            .forEach((field) => {
+              const inputType = getInputType(field, this.options);
+
+              fileContent.push(
+                `${field.name}: ${inputType.isList ? `[${inputType.type}!]` : inputType.type}${
+                  field.isRequired ? '!' : ''
+                }`,
+              );
+            });
+          fileContent.push('}', '');
+        }
+      });
+
+      schema?.outputObjectTypes.prisma
+        .filter((type) => type.name.includes('Aggregate') || type.name.endsWith('CountOutputType'))
+        .forEach((type) => {
+          fileContent.push(`type ${type.name} {`, '');
+          type.fields
+            .filter((field) => !this.options?.excludeInputFields?.includes(field.name))
+            .forEach((field) => {
+              fileContent.push(
+                `${field.name}: ${field.outputType.isList ? `[${field.outputType.type}!]` : field.outputType.type}${
+                  !field.isNullable ? '!' : ''
+                }`,
+              );
+            });
+          fileContent.push('}', '');
+        });
+    }
+    return this.formation(fileContent.join('\n'), 'graphql');
+  }
+
   protected readFile(path: string) {
     return existsSync(path) ? readFileSync(path, { encoding: 'utf-8' }) : '';
   }
 
   protected getImport(content: string, path: string) {
-    return this.isJS
-      ? `const ${content} = require('${path}')`
-      : `import ${content} from '${path}'`;
+    return this.isJS ? `const ${content} = require('${path}')` : `import ${content} from '${path}'`;
   }
 
   protected filterDocs(docs?: string) {
-    return docs
-      ?.replace(/@PrismaSelect.map\(\[(.*?)\]\)/, '')
-      .replace(/@onDelete\((.*?)\)/, '');
+    return docs?.replace(/@PrismaSelect.map\(\[(.*?)\]\)/, '').replace(/@onDelete\((.*?)\)/, '');
   }
 
   protected shouldOmit(docs?: string) {
@@ -227,33 +252,22 @@ export class Generators {
     }
     const innerExpression = docs?.match(/@Pal.omit\(\[(.*?)\]\)/);
     if (innerExpression) {
-      const expressionArguments = innerExpression[1]
-        .replace(/\s/g, '')
-        .split(',')
-        .filter(Boolean);
+      const expressionArguments = innerExpression[1].replace(/\s/g, '').split(',').filter(Boolean);
       return expressionArguments.includes('output');
     }
     return false;
   }
 
-  protected createFileIfNotfound(
-    path: string,
-    fileName: string,
-    content: string,
-  ) {
+  protected createFileIfNotfound(path: string, fileName: string, content: string) {
     !existsSync(path) && this.mkdir(path);
-    !existsSync(join(path, fileName)) &&
-      writeFileSync(join(path, fileName), content);
+    !existsSync(join(path, fileName)) && writeFileSync(join(path, fileName), content);
   }
 
   protected get parser() {
     return this.isJS ? 'babel' : 'babel-ts';
   }
 
-  protected formation(
-    text: string,
-    parser: PrettierOptions['parser'] = this.parser,
-  ) {
+  formation(text: string, parser: PrettierOptions['parser'] = this.parser) {
     return format(text, {
       singleQuote: true,
       semi: false,
